[feat] applying mvp
