+++ app/src/main/AndroidManifest.xml
@@ -13,14 +13,10 @@
             android:roundIcon="@drawable/image_splash_thumbs"
             android:supportsRtl="true"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-        <activity android:name=".ui.setting.SettingActivity">
-        </activity>
-
         <service
                 android:name=".services.ControllerService"
                 android:enabled="true"
                 android:exported="false"/>
-
         <activity
                 android:name=".ui.intro.SplashActivity"
                 android:theme="@style/SplashTheme">
@@ -34,7 +30,7 @@
         </activity>
         <activity android:name=".ui.status.StatusActivity">
         </activity>
-        <activity android:name=".ui.setting.SettingActivity2"
+        <activity android:name=".ui.setting.SettingActivity"
                   android:parentActivityName=".ui.status.StatusActivity">
         </activity>
     </application>
