+++ app/src/main/java/com/thumbs/android/thumbsAndroid/ui/register/RegisterPresenter.kt
@@ -11,13 +11,14 @@ class RegisterPresenter(
 
 
     override fun createThumb(thumbName: String) {
-        if (registerView?.isNotEmptyName() == true)
+        if (registerView?.isNotEmptyName() == true){
+            registerView?.showToast("완료.")
             thumbsRepository.createThumbs(1,  hashMapOf("name" to thumbName)).subscribe({
                 registerView?.showToast("등록이 완료되었습니다.")
                 registerView?.nextPage()
             },{
                 it.printStackTrace()
-            })
+            })}
         else
             registerView?.showToast("이름을 입력해주세요!")
     }
