+++ app/src/main/java/com/thumbs/android/thumbsAndroid/ui/setting/SettingContract.kt
@@ -1,11 +1,12 @@
 package com.thumbs.android.thumbsAndroid.ui.setting
 
+import com.thumbs.android.thumbsAndroid.model.Thumb
 import com.thumbs.android.thumbsAndroid.ui.base.BasePresenter
 import com.thumbs.android.thumbsAndroid.ui.base.BaseView
 
 interface SettingContract {
     interface SettingView : BaseView {
-        fun setUi()
+        fun setUi(thumb: Thumb)
     }
 
     interface SettingUserActionListener : BasePresenter<SettingContract.SettingView> {
