+++ app/src/main/java/com/thumbs/android/thumbsAndroid/ui/setting/SettingPresenter.kt
@@ -1,21 +1,22 @@
 package com.thumbs.android.thumbsAndroid.ui.setting
 
 import android.util.Log
+import com.thumbs.android.thumbsAndroid.repositories.ThumbsRepository
 import com.thumbs.android.thumbsAndroid.repositories.UserRepository
 import com.thumbs.android.thumbsAndroid.ui.base.BasePresenter
 
 class SettingPresenter(
-    val userRepository: UserRepository
+    val thumbsRepository: ThumbsRepository
 ) : SettingContract.SettingUserActionListener {
 
 
     var settingView : SettingContract.SettingView?= null
 
     override fun load() {
-        userRepository.getUser().subscribe ({
-            Log.d("TEST","성공")
+        thumbsRepository.loadThumb(1).subscribe({
+            settingView?.setUi(it)
         },{
-            Log.d("TEST","실패")
+            it.printStackTrace()
         })
     }
