--- app/src/main/res/layout/activity_settings.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  android:layout_height="match_parent"
-  android:layout_width="match_parent"
->
-  <Button
-    android:id="@+id/CreateWidgetButton"
-    android:layout_centerInParent="true"
-    android:layout_height="wrap_content"
-    android:layout_width="wrap_content"
-    android:text="@string/APP_NAME"
-  />
-  <Button
-    android:id="@+id/buttonSetting"
-    android:text="Setting"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_alignParentTop="true"
-    android:layout_marginTop="128dp"
-    android:layout_centerHorizontal="true"/>
-</RelativeLayout>
